requirements.txt
streamlit
langchain-google-genai>=2.0.0
langchain-community
langchain-openai
langchain-pinecone
openai
pinecone-client
docx2txt
python-dotenv
